Multi Zoom Support for Text Layout for win32
This commit adds the support of scaling of text layout based on the zoom
level of the monitor it is drawn on. It uses the native zoom which is
provided by the GCData if available otherwise the zoom level set in the
font (since the right font for a zoom level is computed already at the
control level, so we always have the consistent zoom information).

contributes to #62 and #127