Fix usage of re-scaled fonts in Control:setFont for windows
In Control::setFont the passed font and not the font from the SWTFontProvider is used. This has only an effect when the ScaledFontRegistry is used and a font is passed, that does not match the current zoom of the control.

Contributes to https://github.com/eclipse-platform/eclipse.platform.swt/issues/62
and https://github.com/eclipse-platform/eclipse.platform.swt/issues/127