Fix usage of re-scaled fonts in TableItem::setFont for windows
In TableItem::setFont with the index specified, the passed font and not the font from the SWTFontProvider is used. This has only an effect when the ScaledFontRegistry is used and a font is passed, that does not match the current zoom of the table item.

Contributes to https://github.com/eclipse-platform/eclipse.platform.swt/issues/62
and https://github.com/eclipse-platform/eclipse.platform.swt/issues/127