Remove methods consuming or producing pixel values in Event
In this contribution the DPIUtil dependency from org.eclipse.swt.widgets.Event is removed including all package protected pixel consuming or returning methods. As (at least) Windows receivces HiDPI support for multiple zoom levels, this is a prerequisite to contribute this feature. As a consequence all callers inside SWT that still used the pixel dependent methods are adapt to only use the public API.

Contributes to #62 and #131