Create CTabFolder images with ImageDataProvider
This commit changes the createButtonImage method in CTabFolder. Until now the requested image data were created once and are scaled to other zoom settings with an AutoScaleImageDataProvider, which is a destructive operation when scaling it to a smaller zoom. With this change the whole logic is moved into an ImageDataProvider, so each time ImageData for a specific zoom value is requested, it will be created from scratch for this zoom value.

Contributes to #62 and #131