Program icon scaling using DPI
This commit contributes to proper scaling of Program icon. Since, these
icons are retrieved by OS.SHGetFileInfo which is System DPI-Aware, it
returns the icon with zoom level of the primary monitor at application
startup. Hence, for initialNativeZoom of the Image object to be created for the
icon, the zoom retrieved from the DPI is considered.

contributes to #62 and #127
fixes #1759