MenuItem Image Scaling with DPI aware metrics #62
This commit contributes to scaling of the MenuItem::hBitmap with the DPI
Aware System Metrics for the size of the SM_CYMENUCHECK or the size
obtained by menuItem zoom. It also corrects the zoom in specific cases
where the images are not consistently drawn on the context menu.

contributes to
https://github.com/eclipse-platform/eclipse.platform.swt/issues/62 and
https://github.com/eclipse-platform/eclipse.platform.swt/issues/127